5.1 Рзделение на модули
